webapp/requirements.txt
Flask
psycopg2-binary
python-dotenv
matplotlib
plotly
pandas
holidays
numpy